+++ fabric-screen-handler-api-v1-bridge/src/main/java/dev/loaderbridge/fabric/api/screenhandler/mixin/MenuProviderMixin.java
@@ -1,0 +1,8 @@
+package dev.loaderbridge.fabric.api.screenhandler.mixin;
+
+import net.fabricmc.fabric.api.screenhandler.v1.FabricScreenHandlerFactory;
+import net.minecraft.world.MenuProvider;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(MenuProvider.class)
+public interface MenuProviderMixin extends FabricScreenHandlerFactory { }