+++ fabric-screen-handler-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/screenhandler/v1/FabricScreenHandlerFactory.java
@@ -1,0 +1,8 @@
+package net.fabricmc.fabric.api.screenhandler.v1;
+
+/** Fabric's interface-injected extension for controlling replacement menus. */
+public interface FabricScreenHandlerFactory {
+    default boolean shouldCloseCurrentScreen() {
+        return true;
+    }
+}